requirements.txt
openai>=1.0.0
resend>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
